+++ src/main/java/com/schoolsystem/service/DormitoryService.java
@@ -12,17 +12,19 @@ public interface DormitoryService {
     double getElectricityBalance(int dormitoryId);
 
     // 电费充值
-    void rechargeElectricity(int studentId, int dormitoryId, int amount);
+    void rechargeElectricity(int studentId, int dormitoryId, double amount);
 
     List<DormitoryRecharge> getRechargeHistory(int dormitoryId);
 
     double getWaterBalance(int dormitoryId);
 
-    void rechargeWater(int studentId, int dormitoryId, int amount);
+    void rechargeWater(int studentId, int dormitoryId, double amount);
 
     int addRepair(DormitoryRepair dormitoryRepair);
 
     List<Dormitory> selectAllDormitory();
 
     Dormitory getDormitoryById(Integer id);
+
+
 }
